[REF] outfactor launch file constructer
the launch file constructor is a handy tool for file based workflows,
however in order to workwith data pipelines, the deep integration
was a show stopper.

This commit outfactors the file constructor to its adequat level of abstraction
in the realm of interface scripts (like export and import) and not as part of a
core library item.

Note: The default launchfile creation has been set to fail=False, as this was
causing errors throughout the testing script. It might be better to not cause errores
in the case everything actually worked.